rename pkg to js
