+++ js/pieces/king.ts
