+++ js/pieces/knight.ts
