+++ js/pieces/straight-line.ts
